Enable Korean support
    modified:   base/internal_version.h
    modified:   base/internal_version.h.in
    modified:   base/main.cpp
    modified:   common/str.cpp
    modified:   engines/scumm/actor.cpp
    modified:   engines/scumm/charset.cpp
    modified:   engines/scumm/charset.h
    new file:   engines/scumm/korean.cpp
    new file:   engines/scumm/korean.h
    new file:   engines/scumm/ks_check.cpp
    new file:   engines/scumm/ks_check.h
    modified:   engines/scumm/module.mk
    modified:   engines/scumm/script_v5.cpp
    modified:   engines/scumm/scumm-md5.h
    modified:   engines/scumm/scumm.cpp
    modified:   engines/scumm/scumm.h
    modified:   engines/scumm/smush/smush_player.cpp
    modified:   engines/scumm/string.cpp
    modified:   graphics/font.cpp
    modified:   graphics/font.h
    modified:   graphics/fonts/scummfont.cpp
    new file:   graphics/korfont.cpp
    new file:   graphics/korfont.h
    modified:   graphics/module.mk
    modified:   gui/about.cpp
    modified:   gui/debugger.cpp
    modified:   gui/debugger.h
    modified:   gui/launcher.cpp
    modified:   gui/themes/default.inc
    modified:   gui/themes/scummclassic.zip
    modified:   gui/themes/scummclassic/classic_layout.stx
    modified:   gui/themes/scummclassic/classic_layout_lowres.stx
    modified:   gui/themes/scummmodern.zip
    modified:   gui/themes/scummmodern/scummmodern_layout.stx
    modified:   gui/themes/scummmodern/scummmodern_layout_lowres.stx
    modified:   tools/md5table.c
    modified:   tools/scumm-md5.txt

+++ graphics/font.h
@@ -27,6 +27,7 @@
 
 #include "common/str.h"
 #include "graphics/surface.h"
+#include "graphics/korfont.h"
 
 namespace Common {
 class SeekableReadStream;
@@ -55,8 +56,8 @@ public:
 	virtual int getFontHeight() const = 0;
 	virtual int getMaxCharWidth() const = 0;
 
-	virtual int getCharWidth(byte chr) const = 0;
-	virtual void drawChar(Surface *dst, byte chr, int x, int y, uint32 color) const = 0;
+	virtual int getCharWidth(uint16 chr) const = 0;
+	virtual void drawChar(Surface *dst, uint16 chr, int x, int y, uint32 color) const = 0;
 
 	void drawString(Surface *dst, const Common::String &str, int x, int y, int w, uint32 color, TextAlign align = kTextAlignLeft, int deltax = 0, bool useEllipsis = true) const;
 
@@ -87,8 +88,8 @@ public:
 	virtual int getFontHeight() const { return 8; }
 	virtual int getMaxCharWidth() const { return 8; }
 
-	virtual int getCharWidth(byte chr) const;
-	virtual void drawChar(Surface *dst, byte chr, int x, int y, uint32 color) const;
+	virtual int getCharWidth(uint16 chr) const;
+	virtual void drawChar(Surface *dst, uint16 chr, int x, int y, uint32 color) const;
 };
 
 typedef uint16 bitmap_t; /* bitmap image unit size*/
@@ -132,8 +133,8 @@ public:
 	virtual int getFontHeight() const { return desc.height; }
 	virtual int getMaxCharWidth() const { return desc.maxwidth; }
 
-	virtual int getCharWidth(byte chr) const;
-	virtual void drawChar(Surface *dst, byte chr, int x, int y, uint32 color) const;
+	virtual int getCharWidth(uint16 chr) const;
+	virtual void drawChar(Surface *dst, uint16 chr, int x, int y, uint32 color) const;
 
 	static NewFont *loadFont(Common::SeekableReadStream &stream);
 	static bool cacheFontData(const NewFont &font, const Common::String &filename);
