+++ graphics/fonts/scummfont.cpp
@@ -276,11 +276,11 @@ static const byte guifont[] = {
 	0,   0,   0					// ???
 };
 
-int ScummFont::getCharWidth(byte chr) const {
+int ScummFont::getCharWidth(uint16 chr) const {
 	return guifont[chr+6];
 }
 
-void ScummFont::drawChar(Surface *dst, byte chr, int tx, int ty, uint32 color) const {
+void ScummFont::drawChar(Surface *dst, uint16 chr, int tx, int ty, uint32 color) const {
 	assert(dst != 0);
 	byte *ptr = (byte *)dst->getBasePtr(tx, ty);
